Move command repo to its own file
